+++ src/bin/bat/clap_app.rs
@@ -373,12 +373,63 @@ pub fn build_app(interactive_output: bool) -> Command {
                 .overrides_with("theme")
                 .help("Set the color theme for syntax highlighting.")
                 .long_help(
-                    "Set the theme for syntax highlighting. Use '--list-themes' to \
+                    "Set the theme for syntax highlighting. Note that this option overrides \
+                    '--theme-dark' and '--theme-light'. Use '--list-themes' to \
                      see all available themes. To set a default theme, add the \
                      '--theme=\"...\"' option to the configuration file or export the \
                      BAT_THEME environment variable (e.g.: export \
                      BAT_THEME=\"...\").",
                 ),
+        )
+        .arg(
+            Arg::new("color-scheme")
+                .long("color-scheme")
+                .overrides_with("color-scheme")
+                .value_name("scheme")
+                .value_parser(["auto", "auto:always", "dark", "light", "system"])
+                .default_value("auto")
+                .hide_default_value(true)
+                .help("Specify whether to choose a dark or light theme.")
+                .long_help(
+                    "Specify whether to choose a dark or light syntax highlighting theme. \
+                     Use '--theme-light' and '--theme-dark' (or the environment variables \
+                     BAT_THEME_LIGHT and BAT_THEME_DARK) to configure which themes are picked. \
+                     You may also use '--theme' to set a theme that is used regardless of this choice.\n\n\
+                     Possible values:\n\
+                     * auto (default):\n    \
+                            Query the terminals for its color scheme if the output is not redirected. \
+                            This is to prevent race conditions with pagers such as less.\n\
+                     * 'auto:always':\n    \
+                            Always query the terminal for its color scheme, \
+                            regardless of whether or not the output is redirected.\n\
+                     * dark: Use a dark syntax highlighting theme.\n\
+                     * light: Use a light syntax highlighting theme.\n\
+                     * system: Query the OS for its color scheme. Only works on macOS.\n\
+                     "),
+        )
+        .arg(
+            Arg::new("theme-light")
+                .long("theme-light")
+                .overrides_with("theme-light")
+                .value_name("theme")
+                .help("Sets the color theme for syntax highlighting used for light backgrounds.")
+                .long_help(
+                    "Sets the theme name for syntax highlighting used when the terminal uses a light background. \
+                    Use '--list-themes' to see all available themes. To set a default theme, add the \
+                    '--theme-light=\"...\" option to the configuration file or export the BAT_THEME_LIGHT \
+                    environment variable (e.g. export BAT_THEME_LIGHT=\"...\")."),
+        )
+        .arg(
+            Arg::new("theme-dark")
+                .long("theme-dark")
+                .overrides_with("theme-dark")
+                .value_name("theme")
+                .help("Sets the color theme for syntax highlighting used for dark backgrounds.")
+                .long_help(
+                    "Sets the theme name for syntax highlighting used when the terminal uses a dark background. \
+                    Use '--list-themes' to see all available themes. To set a default theme, add the \
+                    '--theme-dark=\"...\" option to the configuration file or export the BAT_THEME_DARK \
+                    environment variable (e.g. export BAT_THEME_DARK=\"...\")."),
         )
         .arg(
             Arg::new("list-themes")